Fix connection indicator quirks.
Instead of setting the content of the indicator, dynamically, show/hide two fixed elements.
This is important, e.g. if the fail-indicator is an icon served by the embedded server itself.
In this case, it needs to be loaded, *before* the connection is failing, obviously.

Fixes https://github.com/tfry-git/EmbAJAX/issues/19 .